refactor: refactor examples file dir
